+++ src/main/resources/META-INF/org/mule/runtime/core/config/registry-bootstrap.properties
@@ -1,1 +1,1 @@
-extension.objectStore.registryr=org.mule.extension.objectstore.internal.ObjectStoreRegistry
+extension.localObjectStore.registry=org.test.extension.localobjectstore.internal.ObjectStoreRegistry
